fix: Replace default GitLab version and fix for docker-machine download url. (#458)
* feat: Switch to GitLab's forked docker-machine version

Update the version-generated URL to use the GitLab fork of
docker-machine as the official one is dead and buried.

* fix: Default to a version instead of an explicit URL

GitLab has made the "generic" URLs which end in `/docker-machine`
inaccessible, so instead specify a version for docker-machine and use
the URL generation as that both produces URLs that work and should do
the right thing in most cases.

Fixes #457.

* fix: Update runner version to 14.8.2

Version 14.8.0 had a major regression and was removed from all release
distribution channels. Update to 14.8.2 as that's the latest 14.8.x
version.

Fixes #456